refactor(workspace): home openEpicenterRoot in config, drop dead root-barrel exports
The workspace-apps/ folder only opened one Epicenter root, so its plural name and
split files outlived the multi-app model. Merge its WorkspaceAuthClient and
mount-startup errors into open-epicenter-root.ts (matching load-epicenter-config's
inline-errors convention) and move it into config/, which now owns the whole
epicenter.config.ts lifecycle: filename, default source, load, and open.

Also remove three dead exports from the browser-safe root barrel
(DEFAULT_EPICENTER_CONFIG_SOURCE, defineMount, EpicenterRoot); none had a
root-barrel consumer. They stay reachable from @epicenter/workspace/node and
/daemon, and EpicenterRoot moves to the node barrel, its honest home.
